Create WG2meetingATPgeometry
Rule based prover are very interesting in Geometry when we aim to an application in education. Different set of rules are needed for different levels and different national curricula. Preparing a meta-prover capable of accepting different set of rules, formally verified to be consistence and sound, and with natural languages and geometric renderings and capable to be integrated in a dynamic geometry system, is a milestone goal.